Moving client to its own file
