+++ lib/mumukit/auth.rb
@@ -9,6 +9,7 @@ require_relative './auth/slug'
 require_relative './auth/version'
 require_relative './auth/exceptions'
 require_relative './auth/grant'
+require_relative './auth/client'
 require_relative './auth/token'
 require_relative './auth/scope'
 require_relative './auth/permissions'
